+++ discord/utils.py
@@ -92,6 +92,7 @@ __all__ = (
     "format_dt",
     "basic_autocomplete",
     "generate_snowflake",
+    "filter_params",
 )
 
 DISCORD_EPOCH = 1420070400000
@@ -127,7 +128,6 @@ class _cached_property:
 
 
 if TYPE_CHECKING:
-
     from typing_extensions import ParamSpec
 
     from .abc import Snowflake
@@ -1159,3 +1159,35 @@ def basic_autocomplete(values: Values) -> AutocompleteFunc:
         return iter(itertools.islice(gen, 25))
 
     return autocomplete_callback
+
+
+def filter_params(params, **kwargs):
+    """Mainly used in ext.bridge, filter out and replace certain keyword parameters
+
+    Parameters
+    -----------
+    params: Dict[str, Any]
+        The initial parameters to filter.
+    **kwargs: Dict[str, Optional[str]]
+        Key to value pairs where the key's contents would be moved to the
+        value, or if the value is None, remove key's contents (see code example).
+
+    Example
+    -------
+    .. code-block:: python3
+
+        >>> params = {"param1": 12, "param2": 13}
+        >>> filter_params(params, param1="param3", param2=None)
+        {'param3': 12}
+        # values of 'param1' is moved to 'param3'
+        # and values of 'param2' are completely removed.
+
+    """
+    for param, nparam in kwargs.items():
+        if param in params:   
+            if nparam is None:
+                params.pop(param)
+            else:
+                params[nparam] = params.pop(param)
+
+    return params
