+++ docs/api.rst
@@ -1493,6 +1493,8 @@ Utility Functions
 
 .. autofunction:: discord.utils.basic_autocomplete
 
+.. autofunction:: discord.utils.filter_params
+
 .. _discord-api-enums:
 
 Enumerations
