+++ docs/ext/bridge/api.rst
@@ -57,7 +57,6 @@ BridgeCommand
 
 .. autoclass:: discord.ext.bridge.BridgeCommand
     :members:
-    :exclude-members: add_to
 
 .. automethod:: discord.ext.bridge.bridge_command()
     :decorator:
